Bug Fixing
Checked myallocate up until about line 102, everything seems well up to
then.

+++ my_malloc.h
@@ -48,7 +48,7 @@ typedef struct PageNode {
 	int used;
 	/* The size of the current node in bytes.
 	TODO @all: discuss how to use this for the paging project. */
-	unsigned short size; 
+	int size; 
 	
 	//Pointer to the next
 	char * next;
